test: 점수 필드 타입 변경에 따른 테스트 코드 수정

+++ src/test/java/com/part3_team4/deokhoogam/domain/user/PowerUserRankingRepositoryTest.java
@@ -62,7 +62,7 @@ public class PowerUserRankingRepositoryTest {
     PowerUserRanking ranking1 = PowerUserRanking.builder()
         .userId(UUID.randomUUID())
         .period(PowerUserPeriod.DAILY)
-        .score(10.0)
+        .score(BigDecimal.valueOf(10.0))
         .ranking(1)
         .build();
     powerUserRankingRepository.save(ranking1);
